+++ child/src/sys_info.c
@@ -12,6 +12,7 @@
  */
 
 #include "sys_info.h"
+#include <stdio.h>
 
 void
 getPlatformInfo ()